Several adjustments
