+++ src/web3/bob-vue/src/components/AppHeader.vue
@@ -1,9 +1,12 @@
 <template>
   <v-app-bar :elevation="2">
     <template v-slot:prepend>
-      <v-app-bar-title><img src="../assets/bob_logo_horizontal.png" height="52" style="padding: 5px 0 0 0;" /></v-app-bar-title>
       <v-app-bar-nav-icon @click="$emit('toggle-sidebar')"></v-app-bar-nav-icon>
     </template>
+    <v-app-bar-title></v-app-bar-title>
+    <template v-slot:append>
+      <img src="../assets/bob_logo_horizontal.png" height="52" style="padding: 5px 5px 0 0;" />
+    </template>    
   </v-app-bar>
 </template>
