+++ src/web3/bob-vue/src/config.js
@@ -1,5 +1,5 @@
 export default {
     websocketsUrl: 'ws://' + window.location.hostname + ':9090',
-    annotatedTopic: { topic: 'bob/frames/annotated/resized/compressed', type: 'sensor_msgs/msg/CompressedImage' },
-    bgsTopic: { topic: 'bob/frames/foreground_mask/resized/compressed', type: 'sensor_msgs/msg/CompressedImage' },
+    annotatedTopic: { topic: 'bob/frames/annotated/resized/compressed', info: 'bob/camera/all_sky/camera_info' },
+    bgsTopic: { topic: 'bob/frames/foreground_mask/resized/compressed', info: 'bob/camera/all_sky/camera_info' },
   };
