--- src/web3/bob/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-  "recommendations": [
-    "Vue.volar",
-    "dbaeumer.vscode-eslint",
-    "esbenp.prettier-vscode"
-  ]
-}
