--- src/web3/bob/run_dev.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-npm install
-npm run dev
