--- src/web3/bob/src/assets/bob.css
@@ -1,185 +1,0 @@
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-  margin: 0;
-  font-weight: normal;
-}
-
-body {
-  min-height: 100vh;
-  color: var(--p-text-color);
-  background: var(--p-content-background);
-  transition:
-    color 0.5s,
-    background-color 0.5s;
-  line-height: 1.6;
-  font-family:
-    Inter,
-    -apple-system,
-    BlinkMacSystemFont,
-    'Segoe UI',
-    Roboto,
-    Oxygen,
-    Ubuntu,
-    Cantarell,
-    'Fira Sans',
-    'Droid Sans',
-    'Helvetica Neue',
-    sans-serif;
-  font-size: 15px;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-/* @media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-  }
-} */
-
-.bob-app {
-  padding: 5px;
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  width: 100vw;
-  margin: 0;
-  overflow: hidden;
-  font-weight: normal;
-}
-
-.header {
-  flex-shrink: 0;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 20px;
-  background-color: var(--p-mask-background);
-  height: 80px; /* Fixed height for the header */
-}
-
-.logo {
-  font-size: 24px;
-}
-
-.title {
-  font-size: 18px;
-}
-
-.footer {
-  width: 100%;
-  background-color: var(--p-mask-background);
-  text-align: center;
-  padding: 10px 20px;
-  height: 100%; /* Fixed height for the footer */
-  overflow-y: auto;
-}
-
-.features {
-  flex: 1;
-  overflow-y: auto;
-}
-
-.sidebar {
-  flex: 1;
-  overflow-y: auto;
-  /* width: 280px;
-  min-width: 280px;
-  max-width: 280px; */
-  overflow: hidden;
-  background-color: var(--p-content-hover-background);
-  padding: 10px;
-  overflow-y: auto;
-  border-color: rgb(222, 226, 230);
-  border-width: 1;  
-}
-
-.menu {
-  width: 200px;
-}
-
-.menu-item {
-  display: flex;
-  align-items: center;
-  padding: 5px 10px;
-  text-decoration: none;
-  color: #007bff;
-}
-
-.menu-item:hover {
-  background-color: #f8f9fa;
-}
-
-.menu-icon {
-  margin-right: 10px;
-}
-
-.section-title {
-  padding: 5px 10px;
-  font-weight: bold;
-  color: #6c757d;
-}
-
-.main-content {
-  display: flex;
-  justify-content: center;
-  align-items: top;
-  box-shadow: 0 0 3px var(--p-content-borderColor) inset;
-  background-color: var(--p-content-background);
-  height: 100%;
-
-  overflow-y: auto;
-  flex: 4;
-  width: 100%;
-  height: 100%;
-  padding: 0px;
-  overflow-y: auto;
-  background-color: var(--p-content-background);
-}
-
-.imgDisplayDiv {
-  height: 100%;
-}
-.imgDisplay {
-  width: 100%;
-}
-
-.features {
-  flex: 1;
-  overflow-y: auto;
-  height: auto;
-  max-width: 200px;
-  background-color: var(--p-content-hover-background) !important;
-  padding: 20px;
-  overflow-y: auto;
-  min-width: 150px;
-  max-width: 300px;
-}
-.features ul {
-  list-style: none;
-  padding: 0;
-}
-.features li {
-  margin: 10px 0;
-}
-
-.p-splitter {
-  height: 100%;
-  padding: 1px;
-}
-.p-splitterpanel {
-  display: flex;
-  justify-content: center;
-  align-items: top;
-  box-shadow: 0 0 3px var(--p-content-borderColor) inset;
-  background-color: var(--p-content-background);
-  height: 100%;
-}
