--- src/web3/bob/src/assets/main.css
@@ -1,19 +1,0 @@
-@import './base.css';
-
-#app {
-  margin: 0 auto;
-  max-width: 99%;
-  font-weight: normal;
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-  }
-}
