--- src/web3/bob/src/components/Footer.vue
@@ -1,11 +1,0 @@
-<template>
-  <footer>
-    <p>Log display will be here</p>
-  </footer>
-</template>
-
-<script>
-export default {
-  name: 'Footer',
-};
-</script>
