--- src/web3/bob/src/components/MainContent.vue
@@ -1,8 +1,0 @@
-<template>
-  <main>
-    <router-view></router-view>
-  </main>
-</template>
-
-<script>
-</script>
