--- src/web3/bob/src/config.js
@@ -1,3 +1,0 @@
-export default {
-    websocketsUrl: 'ws://' + window.location.hostname + ':9090',
-  };
