--- src/web3/bob/src/views/HomeView.vue
@@ -1,47 +1,0 @@
-<template>
-  <main class="main-content">
-    <div class="imgDisplayDiv">
-      <img ref="imageDisplayWindow" id="imageDisplayWindow" class="imgDisplay" />
-    </div>
-    <Features class="features" />
-  </main>
-</template>
-
-<script>
-import config from '@/config';
-import Features from '../components/Features.vue';
-
-export default {
-  name: 'Annotated View',
-  data() {
-    return {
-      bob: null,
-    };
-  },
-  components: {
-    Features,
-  },
-  mounted() {
-    this.bob = new BobRos(config.websocketsUrl);
-    this.bob.connect(true);
-    this.bob.subscribeTopic('bob/frames/annotated/resized/compressed', 'sensor_msgs/msg/CompressedImage',
-        (message) => {
-          try {
-            const imageElement = this.$refs.imageDisplayWindow;
-            imageElement.src = "data:image/jpeg;base64," + message.data;
-          } catch (error) {
-            // Ignore, lost connection
-          }
-        });
-  },
-  beforeRouteLeave(to, from, next) {
-    this.bob.disconnect();
-    next();
-  },
-  beforeDestroy() {
-    this.bob.disconnect();
-  },
-  methods: {
-  },
-};
-</script>
